+++ openedx2zim/xblocks_extractor/drag_and_drop_v2.py
@@ -20,9 +20,14 @@ class DragAndDropV2(
         # extra vars
         self.content = None
 
-    def download(self, instance_connection):
-        content = instance_connection.get_page(self.xblock_json["student_view_url"])
+    def download_inner(self, instance_connection):
+        url = self.xblock_json["student_view_url"]
+        try:
+            content = instance_connection.get_page(url)
+        except Exception:
+            content = None
         if not content:
+            self.add_failed({"url": url})
             return
         soup = BeautifulSoup(content, "lxml")
         self.content = json.loads(